# back-end/swagger.output.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "E-Commerce API Documentation",
      "version": "1.0.0",
      "description": "API documentation for the e-commerce platform"
    },
    "servers": [
      {
        "url": "http://localhost:3000",
        "description": "Development server"
      }
    ],
    "components": {
      "securitySchemes": {
        "BearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [
      {
        "BearerAuth": []
      }
    ],
...
    "paths": {
      "/auth/register": {
        "post": {
          "tags": ["Auth"],
          "summary": "Register a new user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "firstname": { "type": "string" },
                    "lastname": { "type": "string" },
                    "username": { "type": "string" },
                    "email": { "type": "string" },
                    "password": { "type": "string" },
                    "address": { "type": "string" },
                    "phone": { "type": "string" }
                  }
                }
              }
            }
          },
          "responses": {
            "201": { "description": "User registered successfully" },
            "400": { "description": "Invalid input" },
            "500": { "description": "Internal server error" }
          }
        }
      },
      "/auth/login": {
        "post": {
          "tags": ["Auth"],
          "summary": "Login user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string" },
                    "password": { "type": "string" }
                  }
                }
              }
            }
          },
          "responses": {
            "200": { "description": "User logged in successfully" },
            "400": { "description": "Invalid input" },
            "401": { "description": "Unauthorized" },
            "500": { "description": "Internal server error" }
          }
        }
      },
...
    }
  }